Switch `PLATFORM` to `ruby`
Allows using the Gemfile with any C-based ruby implementation instead of
locking it to the specific CPU architecture of the ruby version that
generated `Gemfile.lock`.

This is habit on Rails apps that are developed on a Mac but deployed on
Linux. It's generally more flexible and avoids adding clutter to the
Gemfile.lock.

The only downside is that gems with native extensions always need to be
built from source, which has been a reasonable trade-off to avoid
frustrating debugging issues or limitations when using `bundler` on a
different CPU architecture.