Set up link to go back to the referring page
Links are added above the `<h1>` and at the end of the `<div>` wrapping
the page content. The links are styled to be hidden when the `href`
attribute is absent.

A `<script>` tag at the end of the `page.html` sets the `href` attribute
to the `return_to` query parameter when
* `document.referrer` points to the Registry
* `return_to` exists and points to the Registry

The `return_to` parameter is cached in a cookie for an hour for
subsequent navigation within the documentation site.